Refactor using a service
