+++ app/controllers/ips_controller.rb
@@ -6,29 +6,20 @@ class IpsController < ApplicationController
     @locations = available_locations
   end
 
-  # rubocop:disable Style/IfInsideElse
   def create
     @ip = Ip.new(create_params)
 
     if @ip.save
-      publish_for_performance_platform
-      publish_radius_whitelist
+      Services::HandleNewIp.new(ip: @ip).execute
       redirect_to(
         ips_path,
-        anchor: 'ips',
         notice: "#{@ip.address} added, it will be active starting tomorrow"
       )
     else
-      if adding_ip_to_location?
-        @location = Location.find_by(id: ip_params[:location_id])
-        render "locations/ips/new"
-      else
-        @locations = available_locations
-        render :new
-      end
+      @locations = available_locations
+      render :new
     end
   end
-  # rubocop:enable Style/IfInsideElse
 
   def index
     set_ip_to_delete if ip_removal_requested?
@@ -51,26 +42,6 @@ private
     end
   end
 
-  def publish_for_performance_platform
-    UseCases::PerformancePlatform::PublishLocationsIps.new(
-      destination_gateway: Gateways::S3.new(
-        bucket: ENV.fetch('S3_PUBLISHED_LOCATIONS_IPS_BUCKET'),
-        key: ENV.fetch('S3_PUBLISHED_LOCATIONS_IPS_OBJECT_KEY')
-      ),
-      source_gateway: Gateways::Ips.new
-    ).execute
-  end
-
-  def publish_radius_whitelist
-    PublishWhitelist.new(
-      destination_gateway: Gateways::S3.new(
-        bucket: ENV.fetch('S3_PUBLISHED_LOCATIONS_IPS_BUCKET'),
-        key: ENV.fetch('S3_WHITELIST_OBJECT_KEY')
-      ),
-      generate_whitelist: UseCases::Radius::GenerateRadiusIpWhitelist.new
-    ).execute
-  end
-
   def ip_params
     params.require(:ip).permit(:address, :location_id, location_attributes: %i[address postcode])
   end
@@ -82,10 +53,6 @@ private
 
   def user_creates_new_location?
     ip_params[:location_id].blank?
-  end
-
-  def adding_ip_to_location?
-    !ip_params.has_key?(:location_attributes)
   end
 
   def params_with_new_location
